更新错误码分配规则。 ZeroWeb 的错误码模块从 -1~-128编号，以方便使用者从 1 开始向后编号。

+++ zeroweb-spring-boot-starter/src/main/java/io/github/xezzon/zeroweb/core/error/IErrorCode.java
@@ -14,8 +14,8 @@ public interface IErrorCode {
     return String.format(
         "%s%02X%02X",
         sourceType().getCode(),
-        Short.toUnsignedInt(moduleCode()),
-        Short.toUnsignedInt(serialNumber())
+        Byte.toUnsignedInt(moduleCode()),
+        Byte.toUnsignedInt(serialNumber())
     );
   }
 
@@ -25,13 +25,17 @@ public interface IErrorCode {
   ErrorSourceType sourceType();
 
   /**
-   * 模块编码 数据范围 1~127、-128~-1 0保留为公共异常 可转为2位16进制数的表示形式
+   * 模块编码
+   * 0 为公共模块
+   * ZeroWeb 的模块编码从 -1~-128 分配
+   * 1~127 由使用者分配
    * @return 模块编码
    */
   byte moduleCode();
 
   /**
-   * 异常码序列号 同一模块下的异常按从1~127、-128~-1编号
+   * 异常码序列号
+   * 同一模块下的异常按从 1~255 分配
    * @return 异常码序列号
    */
   default byte serialNumber() {
